Fixed loading of functions in multiple extensions.
If a function was defined in multiple extensions, it would only be loaded
in the first one that's defined. This could cause some functions to be
missing depending on which are available.

To support this, function pointers are now all cleared at the start of
loading rather than in the else blocks when they are loaded. This ensures
that functions aren't cleared out after previously loading them.

Updated registry version. This required some re-working of type
declarations.