Support choosing loader library at runtime.
Each loader library now has ANYGL_HAS_* macros defined to determine if they
are available. ANYGL_LOAD_DEFAULT is defined to the default library rather
than relying on ANYGL_LOAD for the single supported library. Each internal
function is now decorated so the correct one may be chosen at runtime.

The main use case for this is on Linux, where GLX is used for X11 and EGL
is used for Wayland.